refactored headers, templates, cookies
