+++ HTTPlayground/main.py
@@ -1,9 +1,8 @@
-from sys import argv
 from http.server import HTTPServer, BaseHTTPRequestHandler
-from tools.file_reader import FileReader
-from views import not_found
-from urls import urls
-from settings import URL, PORT
+from HTTPlayground.views import not_found
+from HTTPlayground.urls import urls
+from HTTPlayground.settings import URL, PORT
+from HTTPlayground.base import FileReader, send_headers
 
 
 class SimpleRequestHandler(BaseHTTPRequestHandler):
@@ -13,6 +12,9 @@ class SimpleRequestHandler(BaseHTTPRequestHandler):
             return
         elif self.path.endswith(('.css', '.js')):
             filename = self.path.rsplit('/', 1)[-1]
+            extension = filename.rsplit('.', 1)[-1]
+            content_type_part = "application" if extension == "js" else "text"
+            send_headers(self, content_type=f'{content_type_part}/{extension}')
             self.wfile.write(FileReader.read(filename, mode='b'))
         else:
             body_content = f'{urls.get(self.path, not_found)(self)}'
@@ -34,7 +36,4 @@ def run(server_class=HTTPServer, handler_class=BaseHTTPRequestHandler, url=URL, 
 
 
 if __name__ == '__main__':
-    if len(argv) == 2:
-        run(url=argv[1], port=int(argv[2]))
-    else:
-        run(handler_class=SimpleRequestHandler)
+    run(handler_class=SimpleRequestHandler)
