--- settings.py
@@ -1,8 +1,0 @@
-from pathlib import Path
-
-PORT = 8000
-URL = 'localhost'
-
-BASE_DIR = Path(__file__).parent
-TEMPLATE_DIR = 'templates'
-STATIC_URL = 'static'
