+++ templates/charged.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Charged</title>
+</head>
+<body>
+    <form action="{{link}}" method="get">
+        <p>{{text}}</p>
+        <button type="submit">Home</button>
+    </form>
+</body>
+</html>
