+++ templates/login.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Login</title>
+</head>
+<body>
+    <form action="{{link}}" method="get">
+        <p>Successful login.</p>
+        <button type="submit">Home</button>
+    </form>
+</body>
+</html>
