+++ templates/not_found.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Error 404</title>
+</head>
+<body>
+    <form action="{{link}}" method="get">
+        <p>ERROR 404. Page not found.</p>
+        <button type="submit">Home</button>
+    </form>
+</body>
+</html>
